ci: keep git diff for ci workflow
